requirements.txt
asgiref==3.11.0
Django==6.0
lxml==6.0.2
python-docx==1.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
